requirements.txt
streamlit==1.39.0
Pillow==10.4.0
numpy==2.1.1